requirements.txt
aiohttp==3.6.0
altgraph==0.16.1
apiclient==1.0.4
appdirs==1.4.3
asgiref==3.2.3
async-timeout==3.0.1
attrs==19.1.0
backcall==0.1.0
beautifulsoup4==4.8.0
certifi==2019.6.16
chardet==3.0.4
colorama==0.4.1
configparser==4.0.2
crawler==0.0.2
crayons==0.2.0
decorator==4.4.0
distlib==0.3.0
dj-database-url==0.5.0
Django==3.0.3
django-cors-headers==3.4.0
et-xmlfile==1.0.1
filelock==3.0.12
future==0.17.1
gunicorn==20.0.4
html5lib==1.0.1
httplib2==0.14.0
idna==2.8
image==1.5.31
importlib-metadata==1.5.0
ipython==7.8.0
ipython-genutils==0.2.0
jdcal==1.4.1
jedi==0.15.1
multidict==4.5.2
numpy==1.16.4
openpyxl==3.0.0
pandas==0.25.0
parso==0.5.1
pefile==2019.4.18
pickleshare==0.7.5
Pillow==7.1.2
prompt-toolkit==2.0.9
psycopg2-binary==2.8.5
py2exe==0.9.2.2
Pygments==2.4.2
PyInstaller==3.5
PyMySQL==0.9.3
PyQt5==5.13.1
PyQt5-sip==12.7.0
python-dateutil==2.8.0
pytz==2019.1
pywin32-ctypes==0.2.0
requests==2.22.0
selenium==3.141.0
six==1.12.0
soupsieve==1.9.2
sqlparse==0.3.1
traitlets==4.3.2
urllib3==1.25.3
virtualenv==20.0.7
virtualenvwrapper-win==1.2.6
wcwidth==0.1.7
webdriver-manager==1.8.2
webencodings==0.5.1
whitenoise==5.1.0
yarl==1.3.0
zipp==3.0.0
